backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pillow
ultralytics
pydicom
